Imported Flask_JWT_Extended and tuned the Authentication

+++ resources/songs.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity
 import json
 from models.songs import SongModel
 
@@ -11,22 +11,26 @@ class Song(Resource):
     parser = reqparse.RequestParser()
     parser.add_argument('title', required=True, help="This field cannot be empty!")
     parser.add_argument('artist',required=True,help="This field cannot be empty!")
-    parser.add_argument('genre_id',required=True,help="Song must have genre!")
+    parser.add_argument('genre_id',type=int,required=True,help="Song must have genre!")
 
-    @jwt_required()
+    @jwt_required
     def get(self, _id_):
         song = SongModel.find_by_id(_id_)
         if song:
             return song.json()
         return {'msg': 'Song Not Exists'}
 
+    @jwt_required
     def delete(self,_id_):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'msg': "Admin Previllage Required."}, 401
         song = SongModel.find_by_id(_id_)
         if song:
             song.delete_from_db()
             return {'msg': 'Song deleted successfully!!'}, 200
             #return {'msg': 'Unsuccessful Operation!'}, 400
-        return {'msg', 'Song Not Exists!'}, 400
+        return {'msg': "Song Not Exists!"}, 400
 
     def put(self,_id_):
         request_data = Song.parser.parse_args()
@@ -50,10 +54,18 @@ class Song(Resource):
 
 class SongList(Resource):
     
-    #@jwt_required()
+    @jwt_optional
     def get(self):
-        # same: {'songs': list(map(lambda x: x.json, SongModel.query.all()))}
-        return{'songs': [song.json() for song in SongModel.query.all()]}  # same: SELECT * FROM songs
+        user_id = get_jwt_identity()   # !!! get user_id from JWT Token
+        songs = [song.json() for song in SongModel.query.all()]
+        if user_id: # !!! If user is logged in, show him full data
+            # same: {'songs': list(map(lambda x: x.json, SongModel.query.all()))}
+            return{'songs': songs}, 200  # same: SELECT * FROM songs
+        # !!! If user is not logged in, just give them a few
+        return {
+            'songs' : [song["title"] for song in songs],
+            'msg' : "Log in for more information!!"
+        }, 200
 
     def post(self):
         request_data = Song.parser.parse_args()
